feat(SF-1029): Pass previous props and state to onUpdate and onUpdated (#202)
`onUpdate` and `onUpdated` now receive the previous props and state as
the first and second parameter respectively. The previous `props` and
`state` passed in are shallow clones of the props and state at the end
of the `updated` phase, or the `before-mount` phase when called for the
first time.

The objects passed in are shallow clones. This means that an object on
props or state must be a new object in order to detect changes more than
one level deep.